# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dogly API",
    "description": "Public API V1 (Spec V1 sez. 9). OpenAPI is the mobile client contract.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/dogs/{dog_id}/albums": {
      "get": {
        "tags": [
          "gallery"
        ],
        "summary": "List Albums",
        "operationId": "list_albums_v1_dogs__dog_id__albums_get",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogAlbumListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "gallery"
        ],
        "summary": "Create Album",
        "operationId": "create_album_v1_dogs__dog_id__albums_post",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogAlbumCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogAlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/albums/{album_id}": {
      "get": {
        "tags": [
          "gallery"
        ],
        "summary": "Get Album",
        "operationId": "get_album_v1_albums__album_id__get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Album Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogAlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/albums/{album_id}/photos": {
      "get": {
        "tags": [
          "gallery"
        ],
        "summary": "List Photos",
        "operationId": "list_photos_v1_albums__album_id__photos_get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Album Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogPhotoListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/albums/{album_id}/photos/init": {
      "post": {
        "tags": [
          "gallery"
        ],
        "summary": "Init Photo",
        "operationId": "init_photo_v1_albums__album_id__photos_init_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Album Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogPhotoInitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogPhotoUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/photos/{photo_id}": {
      "patch": {
        "tags": [
          "gallery"
        ],
        "summary": "Update Photo",
        "operationId": "update_photo_v1_photos__photo_id__patch",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogPhotoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogPhotoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gallery"
        ],
        "summary": "Delete Photo",
        "operationId": "delete_photo_v1_photos__photo_id__delete",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dogs/{dog_id}/visibility": {
      "get": {
        "tags": [
          "gallery"
        ],
        "summary": "Get Visibility",
        "operationId": "get_visibility_v1_dogs__dog_id__visibility_get",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogProfileVisibilityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "gallery"
        ],
        "summary": "Put Visibility",
        "operationId": "put_visibility_v1_dogs__dog_id__visibility_put",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogProfileVisibilityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogProfileVisibilityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DogAlbumCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Title"
          },
          "default_visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "PUBLISHED"
            ],
            "title": "Default Visibility",
            "default": "PRIVATE"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DogAlbumCreate"
      },
      "DogAlbumListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DogAlbumOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DogAlbumListResponse"
      },
      "DogAlbumOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover_photo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Photo Id"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count",
            "default": 0
          },
          "default_visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "PUBLISHED"
            ],
            "title": "Default Visibility",
            "default": "PRIVATE"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dog_id",
          "title",
          "created_at"
        ],
        "title": "DogAlbumOut"
      },
...
      "DogPhotoInitRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Content Type",
            "default": "image/jpeg"
          },
          "bytes": {
            "type": "integer",
            "maximum": 15000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Bytes"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 280
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PRIVATE",
                  "PUBLISHED"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "type": "object",
        "required": [
          "bytes"
        ],
        "title": "DogPhotoInitRequest"
      },
      "DogPhotoListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DogPhotoOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DogPhotoListResponse"
      },
      "DogPhotoOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "album_id": {
            "type": "string",
            "title": "Album Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "storage_path": {
            "type": "string",
            "title": "Storage Path"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "PUBLISHED"
            ],
            "title": "Visibility",
            "default": "PRIVATE"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "album_id",
          "dog_id",
          "storage_path",
          "created_at"
        ],
        "title": "DogPhotoOut"
      },
      "DogPhotoUpdate": {
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 280
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PRIVATE",
                  "PUBLISHED"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          }
        },
        "type": "object",
        "title": "DogPhotoUpdate"
      },
      "DogPhotoUploadResponse": {
        "properties": {
          "photo": {
            "$ref": "#/components/schemas/DogPhotoOut"
          },
          "upload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Upload"
          }
        },
        "type": "object",
        "required": [
          "photo",
          "upload"
        ],
        "title": "DogPhotoUploadResponse"
      },
      "DogProfileVisibilityOut": {
        "properties": {
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "PUBLIC"
            ],
            "title": "Visibility",
            "default": "PRIVATE"
          },
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version"
          },
          "consented_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consented At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "public_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Slug"
          },
          "whitelist_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Whitelist Fields"
          }
        },
        "type": "object",
        "required": [
          "dog_id"
        ],
        "title": "DogProfileVisibilityOut"
      },
      "DogProfileVisibilityUpdate": {
        "properties": {
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "PUBLIC"
            ],
            "title": "Visibility"
          },
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version"
          },
          "public_slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Slug"
          },
          "whitelist_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whitelist Fields"
          }
        },
        "type": "object",
        "required": [
          "visibility"
        ],
        "title": "DogProfileVisibilityUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}